Compress epub file with advzip for some nice gains.
We (re-)compress the file both with level 3 and level 4 as the semantics of
these methods have changed with newer versions of advancecomp.

In previous versions 3 and 4 were different "degrees" of 7zip's deflate
algorithm.  Currently, 3 is 7zip's deflate algorithm, while 4 is zopfli's
and there are some cases where 7zip's compresses better and other cases
where zopfli compresses better.

No algorithm seems to be a better compressor than the other, which is why we
use both.